Use inheritance for new authorization class
